implemented LogrepositIngressDataMapper and LogrepositIngressDataMapperTests

+++ src/main/kotlin/com/logreposit/renogyrover/services/logreposit/dtos/ingress/IngressDataModel.kt
@@ -27,7 +27,7 @@ data class FloatField (
 
 data class IntegerField(
         override val name: String,
-        val value: Int
+        val value: Long
 ) : Field(name = name, datatype = DataType.INTEGER)
 
 data class StringField(
